Conflicts:
	src/gurt/telem/telem_consumer.c
	src/gurt/telemetry.c
	src/include/gurt/telemetry.h

+++ src/bio/bio_xstream.c
@@ -512,6 +512,12 @@ is_server_started(void)
 	return nvme_glb.bd_started;
 }
 
+inline d_list_t *
+bio_bdev_list(void)
+{
+	return &nvme_glb.bd_bdevs;
+}
+
 inline bool
 is_init_xstream(struct bio_xs_context *ctxt)
 {
@@ -850,8 +856,12 @@ replace_bio_bdev(struct bio_bdev *old_dev, struct bio_bdev *new_dev)
 	new_dev->bb_tgt_cnt = old_dev->bb_tgt_cnt;
 	old_dev->bb_tgt_cnt = 0;
 
-	d_list_del_init(&old_dev->bb_link);
-	destroy_bio_bdev(old_dev);
+	if (old_dev->bb_removed) {
+		d_list_del_init(&old_dev->bb_link);
+		destroy_bio_bdev(old_dev);
+	} else {
+		old_dev->bb_faulty = true;
+	}
 }
 
 /*
